make BatchNodeIterable.batchIterables accept a long instead of an int

+++ core/src/main/java/org/neo4j/graphalgo/api/GraphAdapter.java
@@ -63,7 +63,7 @@ public abstract class GraphAdapter implements Graph {
     }
 
     @Override
-    public Collection<PrimitiveLongIterable> batchIterables(int batchSize) {
+    public Collection<PrimitiveLongIterable> batchIterables(long batchSize) {
         return graph.batchIterables(batchSize);
     }
